fix np.int deprecated

+++ dset/Sleep-EDF-2018/prepare_sleep-edf-2018.py
@@ -146,7 +146,7 @@ def main():
             duration_epoch = int(duration_sec / epoch_duration)
 
             # Generate sleep stage labels
-            label_epoch = np.ones(duration_epoch, dtype=np.int) * label
+            label_epoch = np.ones(duration_epoch, dtype=np.int32) * label
             labels.append(label_epoch)
 
             total_duration += duration_sec
